* release/1.0.1:
  username not required
  username required and alphanumeric
  create user
  user favorite
  fix favorites
  sidebar admin
  fix column user
  detect android
  android detect
  sort by price and limit 50km
  sort by price and limit 50km
  promotions

+++ app/controllers/api/favorites_controller.rb
@@ -26,9 +26,15 @@ class API::FavoritesController < ApiController
   # DELETE /users/:user_id/favorites/:id
   def destroy
     user = User.find(params[:user_id])
-    favorite = user.favorites.find(params[:id])
-    favorite.destroy
-    head 204
+    if defined?(user)
+      favorite = user.favorites.size > 0 ? user.favorites.find(params[:id]) : nil
+    end
+    if favorite
+      favorite.destroy
+      head 204
+    else
+      render json: { errors: user.errors }, status: 422
+    end
   end
 
   private
